added common.sh for commonly used functions

+++ roboshop/common.sh
@@ -1,5 +1,11 @@
 #!/usr/bin/bash
 
+USER_ID=$(id -u)
+if [ "$USER_ID" -ne 0 ]; then
+  echo -e "\e[31mYou are not a root or sudo user to run this script\e[0m"
+  exit 2
+fi
+
 LOG=/tmp/roboshop.log
 rm -rf $LOG
 
@@ -15,10 +21,4 @@ VALIDATE() {
 PRINT() {
   echo -n -e "$1\t\t..."
 }
-
-USER_ID=$(id -u)
-if [ "$USER_ID" -ne 0 ]; then
-  echo -e "\e[31mYou are not a root or sudo user to run this script\e[0m"
-  exit 2
-fi
